+++ CHANGELOG.md
@@ -42,4 +42,5 @@ All notable, unreleased changes to this project will be documented in this file.
 - Set shipping required as default for product type - #3655 by @jxltom
 - Docker and compose improvements - #3657 by @michaljelonek
 - Fix TypeError on calling get_client_token - #3660 by @michaljelonek
+- Fix countries in Voucher - #3664 by @michaljelonek
 - Make tokenCreate errors return [] when there are no errors - #3666 by @michaljelonek